Update whl_en.md
Please see the following proposed formatting and syntax changes. 

Formatting:
Implement consistent number formatting. 
Example: ### 2. 
Use consistent hashtags and number system 

Inconsistent language list format:
`ch`, `en`, `french`, `german`, `korean`, `japanese`

Choose between abbreviation or long form. 

Table alignment edit. 
Table content edit to lines 336, 345, and 354.
Those three lines all convey the same information relating to text detection, recognition, and classification inference model folder. 

‘The text detection inference model folder. There are two ways to transfer parameters, 1. None: Automatically download the built-in model to `~/.paddleocr/det`; 2. The path of the inference model converted by yourself, the model and params files must be included in the model path’. 

The first sentence fits the parameter of the table. The description of transfer parameters needs to be in a different section. It provides too much information for the table format and makes it difficult to read. 

Suggestion: 

Add extra section after the table: 
‘There are two ways to transfer parameters.
This applies to text detection, recognition, and classification inference model folders.
1. None: Automatically download the built-in model to `~/.paddleocr/det`
2. The path of the inference model converted by yourself, the model and params files must be included in the model path’. 

Syntax clarification:
'Use by code' 

Unclear phrasing. 
Change to: 
'Use code as input'

Syntax change:
Finally, a dictionary ({word_dict_name}.txt) needs to be provided so that when the model is trained, all the characters that appear can be mapped to the dictionary index.

Change to:
Finally, a dictionary ({word_dict_name}.txt) needs to be provided so that when the model is trained, all the character definitions will be mapped to and appear in an accessible index. 


'Output will be a list' 
Change to: 'Output will be generated as a list'. 

'When the built-in model cannot meet ('the' deleted) needs, you need to use your own trained model.'
Question: Who's needs? The end user? Clarify who is being served by the model.